remove enzyme, since it is incompatible with React 18
