+++ package.json
@@ -64,7 +64,6 @@
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^12.1.5",
     "@testing-library/user-event": "^14.2.0",
-    "@types/enzyme": "^3.10.12",
     "@types/jest": "^27.5.1",
     "@types/node": "^17.0.38",
     "@types/react": "^17.0.45",
@@ -76,14 +75,12 @@
     "@types/zxcvbn": "^4.4.1",
     "@typescript-eslint/eslint-plugin": "^5.39.0",
     "@typescript-eslint/parser": "^5.39.0",
-    "@wojtekmaj/enzyme-adapter-react-17": "^0.6.7",
     "autoprefixer": "^10.4.12",
     "babel-jest": "^27.5.1",
     "babel-loader": "^8.2.5",
     "babel-plugin-formatjs": "10.3.11",
     "compression-webpack-plugin": "^10.0.0",
     "css-loader": "^6.7.1",
-    "enzyme": "^3.11.0",
     "es6-promise": "^4.2.8",
     "eslint": "^8.24.0",
     "eslint-plugin-jest-dom": "^4.0.2",
