+++ src/tests/helperFunctions/LoginTestApp-rtl.tsx
@@ -2,9 +2,32 @@ import { render as rtlRender, RenderOptions, RenderResult } from "@testing-libra
 import { ReduxIntlProvider } from "components/ReduxIntl";
 import type { InitialEntry } from "history";
 import { getTestLoginStore, LoginRootState } from "login-init-app";
+import { initialState as loginInitialState } from "login/redux/slices/loginSlice";
+import { initialState as resetPasswordInitialState } from "login/redux/slices/resetPasswordSlice";
 import React from "react";
 import { MemoryRouter } from "react-router-dom";
-import { loginTestState } from "./LoginTestApp";
+import { initialState as configInitialState } from "reducers/LoginConfig";
+
+export const loginTestState: LoginRootState = {
+  config: {
+    ...configInitialState,
+    next_url: "http://localhost/next",
+    mfa_auth_idp: "https://swedenconnect.idp/",
+    eidas_url: "http://eidas.docker",
+    reset_password_url: "/reset-password-url/",
+    csrf_token: "csrf-token",
+    base_url: "/",
+    // default to being in 'configured' state, since only the test of
+    // the splash screen is ever interested in the opposite
+    is_configured: true,
+    debug: true,
+  },
+  login: loginInitialState,
+  app: { is_loaded: true, loading_data: false, request_in_progress: false },
+  notifications: undefined as any,
+  intl: { locale: "en", messages: {} },
+  resetPassword: resetPasswordInitialState,
+};
 
 interface renderArgs {
   state?: Partial<LoginRootState>;
@@ -26,7 +49,5 @@ function render(ui: React.ReactElement, args: renderArgs = {}): RenderResult {
 
 // re-export everything
 export * from "@testing-library/react";
-// mirror some exports from old Enzyme testing helper
-export { loginTestState };
 // override render method
 export { render };
